Implement function to calculate Euclidean distance.
Function get_euclidean_distance() totals the Euclidean distance between
every tile in the problem state hashmap to the goal state hashmap. Now
returns a float.

+++ Problem.h
@@ -34,7 +34,7 @@ class Problem {
         void move_zero_tile(const int &, const int &);
 
         int get_misplaced_tiles() const;
-        int get_euclidean_distance() const;
+        float get_euclidean_distance() const;
 
     private:
         // Create constant int size at compile time
